refactoring of the type coercion logic
