+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/expressions/BaseField.java
@@ -25,7 +25,7 @@ import com.apple.foundationdb.record.ObjectPlanHash;
 import com.apple.foundationdb.record.PlanHashable;
 import com.apple.foundationdb.record.QueryHashable;
 import com.apple.foundationdb.record.metadata.expressions.TupleFieldsHelper;
-import com.apple.foundationdb.record.query.plan.cascades.values.MessageValue;
+import com.apple.foundationdb.record.query.plan.cascades.values.MessageHelpers;
 import com.apple.foundationdb.record.util.HashUtils;
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.MessageOrBuilder;
@@ -49,7 +49,7 @@ public abstract class BaseField implements PlanHashable, QueryComponent {
 
     @Nonnull
     protected Descriptors.FieldDescriptor findFieldDescriptor(@Nonnull MessageOrBuilder message) {
-        return MessageValue.findFieldDescriptorOnMessage(message, fieldName);
+        return MessageHelpers.findFieldDescriptorOnMessage(message, fieldName);
     }
 
     @Nullable
@@ -57,7 +57,7 @@ public abstract class BaseField implements PlanHashable, QueryComponent {
         if (message == null) {
             return null;
         }
-        return MessageValue.getFieldOnMessage(message, fieldName);
+        return MessageHelpers.getFieldOnMessage(message, fieldName);
     }
 
     @Nonnull
