+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/NullableArrayTypeUtils.java
@@ -23,7 +23,7 @@ package com.apple.foundationdb.record.query.plan.cascades;
 import com.apple.foundationdb.record.RecordMetaDataProto;
 import com.apple.foundationdb.record.metadata.expressions.NestingKeyExpression;
 import com.apple.foundationdb.record.query.plan.cascades.typing.Type;
-import com.apple.foundationdb.record.query.plan.cascades.values.MessageValue;
+import com.apple.foundationdb.record.query.plan.cascades.values.MessageHelpers;
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.Message;
 
@@ -106,7 +106,7 @@ public class NullableArrayTypeUtils {
         // wrapper.
         //
         if (wrappedValue != null && type.getTypeCode() == Type.TypeCode.ARRAY && type.isNullable()) {
-            return MessageValue.getFieldOnMessage((Message)wrappedValue, NullableArrayTypeUtils.getRepeatedFieldName());
+            return MessageHelpers.getFieldOnMessage((Message)wrappedValue, NullableArrayTypeUtils.getRepeatedFieldName());
         }
         return wrappedValue;
     }
