+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/SemanticException.java
@@ -40,12 +40,15 @@ public class SemanticException extends RecordCoreException {
         // generic
         INCOMPATIBLE_TYPE(1, "A value cannot be assigned to a variable because the type of the value does not match the type of the variable and cannot be promoted to the type of the variable."),
         FIELD_ACCESS_INPUT_NON_RECORD_TYPE(3, "A field is accessed on an input that is not of a record type."),
-        RECORD_DOES_NOT_CONTAIN_FIELD(2, "A non-existing field is accessed in a record."),
-        COMPARAND_TO_COMPARISON_IS_OF_COMPLEX_TYPE(3, "The comparand to a comparison expecting an argument of a primitive type, is invoked with an argument of a complex type, e.g. an array or a record."),
-        ARGUMENT_TO_ARITHMETIC_OPERATOR_IS_OF_COMPLEX_TYPE(3, "The argument to an arithmetic operator expecting an argument of a primitive type, is invoked with an argument of a complex type, e.g. an array or a record."),
+        RECORD_DOES_NOT_CONTAIN_FIELD(4, "A non-existing field is accessed in a record."),
+        COMPARAND_TO_COMPARISON_IS_OF_COMPLEX_TYPE(5, "The comparand to a comparison expecting an argument of a primitive type, is invoked with an argument of a complex type, e.g. an array or a record."),
+        ARGUMENT_TO_ARITHMETIC_OPERATOR_IS_OF_COMPLEX_TYPE(6, "The argument to an arithmetic operator expecting an argument of a primitive type, is invoked with an argument of a complex type, e.g. an array or a record."),
 
         // insert, update, deletes
-        UPDATE_TRANSFORM_AMBIGUOUS(1000, "The transformations used in an UPDATE statement are ambiguous.");
+        UPDATE_TRANSFORM_AMBIGUOUS(1_000, "The transformations used in an UPDATE statement are ambiguous."),
+
+        //
+        UNSUPPORTED(10_000, "The action is currently unsupported");
 
         final int numericCode;
         final String message;
